Сохраняет прогресс по работе над попапом, подгрузке комментариев и рефакторингу кода
