+++ src/view/movie-details-view.js
@@ -1,201 +1,5 @@
 import { createElement } from '../render.js';
-import { humanizeDayDate, adaptCommentDate, getTimeFromMins } from '../utils.js';
-
-const createCommentTemplate = (commentItem) => {
-  const {
-    author,
-    comment,
-    commentDate,
-    emotion,
-  } = commentItem;
-
-  const getAdaptedCommentDate = (date) => {
-    if (date) {
-      return adaptCommentDate(date);
-    }
-
-    return '';
-  };
-
-  return `<li class="film-details__comment">
-            <span class="film-details__comment-emoji">
-              <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-sleeping">
-            </span>
-            <div>
-              <p class="film-details__comment-text">${comment}</p>
-              <p class="film-details__comment-info">
-                <span class="film-details__comment-author">${author}</span>
-                <span class="film-details__comment-day">${getAdaptedCommentDate(commentDate)}</span>
-                <button class="film-details__comment-delete">Delete</button>
-              </p>
-            </div>
-          </li>`;
-};
-
-const createAllComments = (movieComments) => {
-  let commentsList = '';
-
-  for (const comment of movieComments) {
-    if (comment) {
-      commentsList = commentsList + createCommentTemplate(comment);
-    }
-  }
-
-  return commentsList;
-};
-
-const createMovieDetailsTemplate = (movie, movieComments) => {
-  const {
-    comments,
-    filmInfo: {
-      title,
-      alternativeTitle,
-      totalRating,
-      poster,
-      ageRating,
-      director,
-      writers,
-      actors,
-      release: {
-        date,
-        releaseCountry
-      },
-      runtime,
-      genre,
-      description,
-    },
-    userDetails: {
-      watchlist,
-      alreadyWatched,
-      favorite,
-    }
-  } = movie;
-
-  const isWatchlistActive = watchlist ? 'film-details__control-button--active' : '';
-  const isAlreadyWatchedActive = alreadyWatched ? 'film-details__control-button--active' : '';
-  const isFavoriteActive = favorite ? 'film-details__control-button--active' : '';
-
-  const getHumanizeDayDate = (dayDate) => {
-    if (dayDate) {
-      return humanizeDayDate(dayDate);
-    }
-    return '';
-  };
-
-  const getTimeInHoursAndMins = (timeInMinutes) => getTimeFromMins(timeInMinutes);
-
-  return (
-    `<section class="film-details">
-    <form class="film-details__inner" action="" method="get">
-      <div class="film-details__top-container">
-        <div class="film-details__close">
-          <button class="film-details__close-btn" type="button">close</button>
-        </div>
-        <div class="film-details__info-wrap">
-          <div class="film-details__poster">
-            <img class="film-details__poster-img" src="${poster}" alt="${title}">
-            <p class="film-details__age">${ageRating ? `${ageRating}+` : ''}</p>
-          </div>
-
-          <div class="film-details__info">
-            <div class="film-details__info-head">
-              <div class="film-details__title-wrap">
-                <h3 class="film-details__title">${title}</h3>
-                <p class="film-details__title-original">${alternativeTitle || ''}</p>
-              </div>
-
-              <div class="film-details__rating">
-                <p class="film-details__total-rating">${totalRating}</p>
-              </div>
-            </div>
-
-            <table class="film-details__table">
-              <tr class="film-details__row">
-                <td class="film-details__term">Director</td>
-                <td class="film-details__cell">${director}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Writers</td>
-                <td class="film-details__cell">${writers}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Actors</td>
-                <td class="film-details__cell">${actors}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Release Date</td>
-                <td class="film-details__cell">${getHumanizeDayDate(date)}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Runtime</td>
-                <td class="film-details__cell">${getTimeInHoursAndMins(runtime)}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Country</td>
-                <td class="film-details__cell">${releaseCountry}</td>
-              </tr>
-              <tr class="film-details__row">
-                <td class="film-details__term">Genres</td>
-                <td class="film-details__cell">
-                  ${genre.map((genreItem) => `<span class="film-details__genre">${genreItem}</span>`).join('')}
-                </td>
-              </tr>
-            </table>
-
-            <p class="film-details__film-description">${description}</p>
-          </div>
-        </div>
-
-        <section class="film-details__controls">
-          <button type="button" class="film-details__control-button film-details__control-button--watchlist ${isWatchlistActive}" id="watchlist" name="watchlist">Add to watchlist</button>
-          <button type="button" class="film-details__control-button film-details__control-button--watched ${isAlreadyWatchedActive}" id="watched" name="watched">Already watched</button>
-          <button type="button" class="film-details__control-button film-details__control-button--favorite ${isFavoriteActive}" id="favorite" name="favorite">Add to favorites</button>
-        </section>
-      </div>
-
-      <div class="film-details__bottom-container">
-        <section class="film-details__comments-wrap">
-          <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-          <ul class="film-details__comments-list">
-          ${createAllComments(movieComments)}
-          </ul>
-
-          <div class="film-details__new-comment">
-            <div class="film-details__add-emoji-label"></div>
-
-            <label class="film-details__comment-label">
-              <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-            </label>
-
-            <div class="film-details__emoji-list">
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-              <label class="film-details__emoji-label" for="emoji-smile">
-                <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-              <label class="film-details__emoji-label" for="emoji-sleeping">
-                <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-              <label class="film-details__emoji-label" for="emoji-puke">
-                <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-              </label>
-
-              <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-              <label class="film-details__emoji-label" for="emoji-angry">
-                <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-              </label>
-            </div>
-          </div>
-        </section>
-      </div>
-    </form>
-  </section>`
-  );
-};
+import { createMovieDetailsTemplate } from '../templates/create-movie-details-template.js';
 
 export default class MovieDetailsView {
   #element = null;
