+++ src/view/no-movies-view.js
@@ -1,0 +1,29 @@
+import { createElement } from '../render.js';
+
+const createNoMoviesTemplate = () => (
+  `<section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title">There are no movies in our database</h2>
+    </section>
+  </section>`
+);
+
+export default class NoMoviesView {
+  #element = null;
+
+  get template() {
+    return createNoMoviesTemplate();
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
